Fixed the issue where closed connection exception thrown from AWS CRT sync HTTP client was not retried.
